readers/trajectory.lua: load simulation parameters early in the script.
The H5MD file needs to be opened before loading mdsim modules,
as readers.h5md retrieves the simulation parameters from the
parameter group within the file. However, the reader requires an
allocated phase space sample with the correct number of particle
species. To disentangle these requirements, we delay the connection
of the phase space sample to reader, through encapsulation in a
property reader.sample. If the position and/or velocity module
accesses reader.sample, the sample's data slots will be connected to
the reader. We cache the sample in a variable outside the property
to ensure that this happens only once.