nose_hoover: follow global changes in master branch
Integrators are connected to mdsim::core via signals.

The timestep is obtained from mdsim::clock.

A module logger is introduced.

The profiling timers are organised differently.

Module dependencies are stored as private shared_ptrs with possible const
attribute.

Connection to the writers for output of the chain variables and the chain
energy is not yet working.