script: support yielding from HALMD script.
This allows interrupting the setup() function with a sampler run,
and continue thereafter, e.g. to thermalize the system in a first
sampler run, and then add observables in a second sampler run.

Note that luabind::resume_function and luabind::resume do not handle
errors properly, but instead abort with an assertion error, so we
have to resort to a solution using the Lua C API in a future commit.