logger: always filter log messages below info severity in Release mode.
When compiling in Release mode and thus with -DNDEBUG, the LOG_DEBUG
macro evaluates to an empty statement. If HALMD_LOG is used directly,
however, e.g. in profiler after commit c64cb00, we have to ensure log
messages with debug severity or lower are filtered regardless of the
severity level set with the --verbose option.