script: use Lua C API for script coroutine.
This adds proper error handling with Lua stack traces.

While the simulation is setup with Lua scripting, it entirely runs in
C++, outside of the Lua interpreter. This avoids unnecessary Lua errors
in case a C++ exception is thrown inside a module, e.g. if a defect GPU
fails during a long simulation run and a CUDA error is thrown. Further
it simplifies stack tracebacks when running inside a debugger.

To allow the script to run multiple partial simulations, we use Lua
coroutines. The Lua function script() is started, and executes until
returning or yielding a slot, e.g. sampler.run. The slot is then
executed in outside of the Lua interpreter. If the function has
yielded, it is then resumed. The Lua function finishes on return
of the last slot.