examples/packages.mk: add Luabind patch for cast graph cache bug.
This patch fixes a severe bug in the Luabind cast graph cache, which
manifests itself with ambiguous overload errors or segmentation faults
due to casts of a Lua value to a mismatching C++ type. However, it
seems to appear only on the x86 platform, x86_64 seems unaffected.

See the description in the patch file for details:

http://sourceforge.net/projects/halmd/files/patches/luabind_cast_graph_cache_invalid_cast_fix.patch