gpu::samples::phase_space: add position and velocity functors for debugging.
Add getter methods which copy particle positions or velocities from the
GPU to the host, along with wrapping functors for use with the H5MD
writer. This allows direct inspection when debugging.

As these methods should never be used in a production run with regard to
the inefficient copying, add #ifndef NDEBUG guards to prevent accidental
usage.