dynamics: refactor Lua modules of correlation functions.
The Lua modules are designed such that a single module initializes and
returns a single C++ module object. This design was violated by the
correlation module, which created multiple correlation functions.
This broke use of other modules such as the module profiler.

Add each correlation function separately in the HALMD script.
At a later stage, we will add the possibility to select a subset
of observables (not restricted to correlation functions).

Store tcf as property of correlation for profiler description.

Rename name() to description() in the correlation function modules.
Remove dead code for H5MD parameters.