Remove global --time option to work around bug in options implementation.
The options parser is currently incapable of assigning global options
only to the module that defines them. In this particular case, the
global --time option sets args.time for all modules. This overrides
args.steps in the args table of the trajectory reader.

This fixes the trajectory continuation test broken in commit 2b352ef1.

This is only a work-around until a complete rewrite of option parsing.