provide pair_full force algorithm
The algorithm naïvely computes N² interactions. This is a first step towards
long-ranged interactions, although more sophisticated (but approximate)
algorithms exist. On the GPU, the scaling is almost linear for small particle
numbers as an increasing fraction of the device is used.

The algorithm can not yet be selected from the command line.

The GPU implementation does not use any tiling algorithm as their impact on
runtime performance is negligible (or may even be negative), see B.Sc. thesis
of Michael Kopp (University of Stuttgart, 2011).