phase_space: remove separation by particle species
Instead of multiple position and velocity arrays, with one each for each
species, define a single position and velocity array in phase_space.

Add a type array for the particle species.

The particle tags are now continuous across all particles of all
species, *not* per species as before.

We will later introduce groups, which will allow the selection of
particles by species, or other criteria. This will restore support
for the calculation of correlation functions or other properties
of a subset of particle, e.g. a single species.