lennard_jones_simple: support more than one species with ε=1, σ=1
Expect scalar value for args.cutoff if ε=1, σ=1 for all species.

In the case of a complex system with more than one particle instance,
it may be useful to have multiple species with the same potential
parameters ε=1 and σ=1.

An example is a system with dimer particles {N, C} in solvent {A, B},
where the interaction parameters for AA, AB and BB are ε=1, σ=1.