potentials: support two particle instances and arbitrarily many species
With two different particle instances, the potential parameters
can no longer be expressed using a symmetric matrix, instead a
full matrix is needed.

Pass the potential parameters as a (full) matrix instead of 3-element
array, which was a remainder from the simulation of binary mixtures.

Allow passing potential parameters to Lua modules either as matrix
(Lua table with tables as rows), or as a scalar value. For a scalar
value, create a matrix and fill the matrix with this value.