maximum_squared_displacement: support multiple module instances
The CUDA kernel was using a module __constant__ for the number of
particles, which varies between particle instances. As the kernel
constant was initialised only once, at construction of a module
object, the number of particles was overwritten by subsequently
constructed objects.

As kernel arguments are stored in constant memory on the Fermi
architecture, kernel constants no longer provide any benefit.
Therefore replace the kernel constant with a kernel argument.