singleton.lua: fix singleton modules with Lua 5.2
For unknown reasons the deprecated module() function in Lua 5.2 did not
invoke the __call metamethod of halmd.singleton, which resulted in all
singleton modules creating multiple instances.

Convert halmd.singleton to a “new style” module, i.e. do not use the
module() command, and explicitly return the register() function. This
function is returned to the user upon including the halmd.singleton
module with require().

Note that without module(), a module is *not* set in the global
namespace, upon require() it must be captured to a local variable.