fix broken ssf unit test
The reference values were wrong in one case as reported by the test. The test
has been revised such that the reference values are computed directly via rules
on the (hkl) indices of the wavevectors. This offers more flexibility and
promises a better maintainability of the test.

This commit fixes #76.