adjust unit test random/distributions
The stochastic test failed regularly.

The limits were set to 3σ, with corresponds to 99.7% probability for a single
test to pass. But since a single test run covers many test cases, the
probability that one of the 18 test cases fails, is about 5%. For several daily
checks of the whole test suite on different machines, a regular failing of the
test suite is likely.

Measures:
1) Increase the confidence interval to 4.5σ which is expected to contain the
result with in 99.999% of the cases.

2) Do not test very small sample sets (n=100, 1000), since the estimator
function for the sum of uniformly distributed variables is not yet a Gaussian.

In essence, 10000 passes of the revised test suite did not fail.

Add more verbose test output.

This commit refs #5.