halmd::fixed_pow: use template recursion
This yields much more efficient code with GCC 4.4 than the implementation with
a while loop, although the loop conditions were known at compile time.

Extend unit test to cover halmd::pow as well. It appears that std::pow calls
libm for variable integer exponent, so the specialised algorithm is preferred
on the host as well.