adopt time correlation functions to particle groups
Pass a list of samplers (phase_space or density_mode) to the correlation
module. The TCF functor will be constructed depending on the memory location of
the underlying samples by overloaded constructors. Thus, the backend for the
computation is selected at the level of the user script. It further allows to
share samplers among different correlation functions.

There is a nasty issue with luabind::property when attaching the functor
description.