fixed_vector: provide comparison operator
Although floating-point comparisons are not advocated, all operators shall be
defined for completeness. Further, the class may be parametrised with an
integral type.

At this point, two fixed_vectors can only be tested for equality. The
definition of the complement, operator!=, is ambiguous. Two vectors may be
unequal if they differ in _all_ elements or in _any_ element. This requires a
more complicated interface, which is postponed for the moment. See the C++
library `Eigen' for an example:

http://eigen.tuxfamily.org/dox/TutorialReductionsVisitorsBroadcasting.html#TutorialReductionsVisitorsBroadcastingReductionsBool