Build loadable HALMD GPU plugin for GPU-agnostic HALMD executable
So far we have been building a library libhalmd.so with both non-GPU and
GPU modules, as well as an executable halmd, which links against this
library. This is not a good solution, since libhalmd.so was never
intended as an end-user library (with a stable API!), rather as an
internal library that unit tests link against.

We continue to build libhalmd.so, but strictly for use with unit tests.
The library is not installed on “make install” any longer.

Further the modules are split into a collection of base modules, which
are statically linked against the halmd executable, and a collection of
GPU modules bundled in a dynamically loadable module gpu.so.

In device.lua, gpu.so is loaded using a protected call of require(),
which ensures that HALMD continues without GPU support if the GPU
module fails to load, e.g. on a system without CUDA libraries.

This way the HALMD executable runs both on GPU and non-GPU platforms.