particle_group: move selection by species from C++ to Lua
The selection by species makes two assumptions: The particles of the
same species have a consecutive tag range, and the number of particles
per species is constant. This is satisfied by a bulk Lennard-Jones
liquid, but not by a biochemical system with complex objects and
catalytic reactions.

Move the C++ code for selection by species to Lua, as the C++ code
should implement the general case. This change removes the duplicate
constructors in particle_group. Further this eliminates the dependency
on particle::ntypes, which will be deprecated soon, as justified above.