program_options.lua: rename optional function “notifier” to “action”
The name “action”, as used in Python's argparse, is more appropriate to
the use of the argument handler: To process the parsed argument value
using a custom function, and store the result in the args table.

For consistency, do *not* store the value if an “action” is specified,
i.e. the default action is to store the value in the args table.