program_options: pass table to notifier functions
In the Lua wrapper of po::notify, push a table onto the stack before
invoking po::notify. In the notifier functors, copy the table from the
stack and pass it to the notifier function. This allows the notifiers to
set argument values without misuse of a class variable in parse_args().