posix_signal: fix signal handling when GPU is enabled
The CUDA library interfers with signal blocking.

We can only speculate as to what is happening in libcuda.so:

   # strings /usr/lib64/libcuda.so | grep ^sig
   sigemptyset
   sigaddset
   sigprocmask

If we block signals before creating a CUDA context in the device module,
signal handling works as expected. If we block signals after creating a
CUDA context, signal handling fails, e.g. on TERM the process aborts.

As a work-around, we block *all* signals using static initialisation,
and register the handlers later, upon construction of posix_signal.

The static variable posix_signal::set_ is used by class members, which
prevents the compiler from optimizing away the call to block_signals().