posix_signal: ignore unhandled signals
This restores the behaviour before commit 9b330d2, i.e. to only notify
of signals with a handler. We register handlers for signals where the
user expects an action (TERM, INT), so it is not necessary to inform
the user of other signals without a default action (e.g. WINCH, sent
on console window resize).