writers.h5md: remove deprecated writing of halmd parameters
Analogous to the options, there no longer exists a set of predefined
parameters for the modules. We will need to reimplement this feature
using a convenient interface to write HDF5 attributes, similar to
halmd.utility.program_options. This will allow the user to store a
custom set of parameters of the simulation.