writers::h5md: create dataset upon first write
Prior to this commit, the dataset was created upon calling on_write,
which required a premature call to the passed data slot(). This is
fine if the data slot merely returns a reference to a member variable
of a data sample. However, if the data slot triggers the sampling
itself, this will cause premature sampling, which is problematic
since the source data, e.g. particle, is likely uninitialised.

In the writer slot, use H5::DataSet::getId() to check if the dataset
of a writer is valid, and if not, create the dataset.

Revert to static functions in the source file again, to avoid symbols.