Revise static structure factor and density modes.
This commit combines three entangled major changes.

We decouple the ssf and observables::{gpu, host}::density_mode modules.
Prior to this commit, the density_mode modules would construct a sample
themselves, which resulted in a strong coupling between the ssf and
the density_mode modules, and the necessity of a base class.

Since we define a separate data structure samples::density_mode, the
ssf and density_mode modules should *not* couple directly, but using the
sample.

For the exchange of the sample itself, adopt a new approach: Pass the
pair of samples *directly* to ssf::sample, and return a const-reference
to the result data structure. This results in a more straight-forward
C++ API, i.e. input data is passed into the function, and outut data is
returned.  For the Lua bindings, wrap ssf::sample using input and output
data slots.

Revise the ssf module by removing the outer containers for handling
multiple pairs of density modes samples. This results in a *much*
simpler C++ code, and also reflects the testing in the unit test.
The normalisation factor, equal to the *total* number of particles,
can be calculated in Lua, by summing over the size of all particle
groups used to calculate the static structure factor.

Remove all class methods which wrap attributes of other modules,
e.g. density_mode::wavenumber wrapping wavevector::wavenumber.
Instead, return the module itself, in density_mode::wavevector.

The two latter changes follow a core design principle of HALMD:
Write modules that do one thing but do it well.

Remove German spelling in comments of gpu::density_mode members.

Declare internal class types as private.

Move profiling typedefs and runtime struct to end of class definition.