samples::blocking_scheme: use a sample slot to acquire sample
In commits d1a8f7e and f6a429a, we revised the implicit copy constructors
of density_mode and phase_space samples to perform a deep instead of a
shallow copy.

Therefore, construct blocking_scheme with a input sample slot instead of
a sample pointer, and in blocking_scheme::push_back acquire a sample by
invoking this slot.

In passing, this commit fixes an issue with dynamics::blocking_scheme:
in dynamics::blocking_scheme::sample, samples are now only acquired if
needed.