signal: use preferred syntax boost::function of Boost.Function
Boost.Function has a preferred and a portable syntax, where the latter
is supported by some older compilers for the Windows platform and the
former is not.

http://www.boost.org/doc/libs/release/doc/html/function/tutorial.html

Since we use a function template parameter for halmd::signal anyway, use
the preferred syntax of Boost.Function. This way boost::function is used
both for signal slots and data slots.