mdsim.integrators: implement connect() and disconnect() methods
Move the connection of integrator slots to an object method connect(),
and collect the returned signal connections in a table. Add an object
method disconnect(), which disconnects all connections in that table.

The code is quite repetitive, however it is better to have properly
working disconnection of integrators now, instead of an abstraction
for all modules designed in a hurry.

This commit fixes #79.