samples/blocking_scheme: explicitly use shared_ptr<sample_type const> for sample
A reader of the code should be able to see at once that the slot returns
a shallow copy (i.e. shared_ptr) of the sample data, not a deep copy.

This is consistent with the rest of the HALMD code, where we never use
typedefs of boost::shared_ptr<T>, but always explicitly state the
pointer container.