observables.phase_space: merge samplers to gpu and host into single Lua object
Create a Lua interface that combines both phase_space sampling to gpu
and host memory in a common object. This allows device-agnostic user
scripts: When running on the host, only a host→host sampler is needed,
while, when running on the gpu, gpu→gpu and gpu→host samplers are
needed.

Move the “memory” argument to the Lua class method acquire().

Add connect() and disconnect() methods to phase_space sampler.

Add connect() and disconnect() methods to phase_space writer.