Move particle_group from namespace observables to mdsim.
The particle_group data structure is complementary to particle, and
defines a subset of particles in particle, whether used for sampling
observables, or to select a subset of particles for integration,
building of neighbour lists, … which we may implement in the future.

Merge libhalmd_luaopen_* functions into one function per translation unit.