samples/phase_space: cache phase_space sample in position/velocity/species slot
The position/velocity/species slot returns a (const) reference to an
array, and thus depends on the sample being alive upon use of the array.

This is guaranteed if the sample is cached in the observable, which is
the case for phase_space::acquire(), however, we should not depend on
this caching. Therefore cache sample in the position/velocity/species
slot itself.