Merge initialisation of particles from phase_space sample.
The separation into copying of positions and types from a phase_space
sample, and copying of velocities from a phase_space sample, will make
it difficult to support additional (optional) particle arrays, i.e.
orientations.

Further the separation is not consistent with phase_space::acquire().

Merge phase_space::set() methods into a common method in phase_space
sampler, since there is sufficient overlap of dependencies and array
buffers.

This is only a preliminary implementation, which lacks flags for partial
sampling (e.g. without velocities), and toggling of the initialisation
of image vectors (for resuming from trajectory file with image vectors
set to zero). Further the execution dimensions of GPU kernels and the
size of the array buffers do not match the particle group size yet.