pair_full_kernel: fix FTBFS with CUDA 3.1
Since commit f1e8795, nvcc 3.1 fails to compile variant constants and
textures. This is due to a bug in C++ namespace handling in nvcc 3.1,
which appears to have been fixed in nvcc 4.1 or earlier.

Remove use of variant constant in pair_full kernel, and pass random
number generator as a kernel argument instead. This also pursues the
long-term goal of removing all module-level kernel constants.