neighbours/phase_space_kernel: fix FTBFS with CUDA 3.1
Since commit f1e8795, nvcc 3.1 fails to compile variant constants and
textures. This is due to a bug in C++ namespace handling in nvcc 3.1,
which appears to have been fixed in nvcc 4.1 or earlier.

Remove use of variant constant in phase_space kernel, and pass
constants as a kernel arguments instead. This also pursues the
long-term goal of removing all module-level kernel constants.

Define a texture<float4> for minimum image vectors, and reinterpret_cast
to the texture<float4> or texture<float2> depending on the dimension.
The data type of the texture is irrelevant, as long as we bind the
correct array to the texture before kernel invocation.