neighbours/particle_kernel: fix FTBFS with CUDA 3.1
Since commit f1e8795, nvcc 3.1 fails to compile variant constants and
textures. This is due to a bug in C++ namespace handling in nvcc 3.1,
which appears to have been fixed in nvcc 4.1 or earlier.

Define a texture<float4> for minimum image vectors, and reinterpret_cast
to the texture<float4> or texture<float2> depending on the dimension.
The data type of the texture is irrelevant, as long as we bind the
correct array to the texture before kernel invocation.