box: remove initialisation using number of particles and density
This removes the dependency of box on the number of particles, which
otherwise prevents construction of box before particle instances.

In the general case, the box must be available before the particle
instances, as the number of particles might be derived from the box
extents, e.g. when filling a system of large spheres with particles
on a lattice.

Remove box::density().

Adapt unit tests, all of which simulate a bulk fluid.