numeric: add compatibility wrappers for packing of fixed_vector<float, …>
This allows using the <<= operator to convert between a tuple of
fixed_vector<float, …> and scalar to a *tuple* of two float4,
which avoids macros or duplicate code to support float, dsfloat
and double with the same kernel.

Note that we pass the second, unneeded float4 value as a reference
(if on the left) or const reference (if on the right). Since the
reference is never used, the memory will *not* be accessed.