verlet_kernel: revert to using USE_VERLET_DSFUN to differ between single/double
While discarding of float4 by reference in the case of single precision
works fine, we also need to support float4 by *value* for tex1Dfetch.
This would *not* be compliant with the C++ standard, since the call
to tex1Dfetch would only be omitted at the mercy of the compiler's
optimizer.

Thus revert to separate code for single/double precision to CUDA vector conversion.