numeric: remove discard of second float4 for single precision to CUDA vector converter
While discarding of float4 by reference in the case of single precision
works fine, we also need to support float4 by *value* for tex1Dfetch.
This would *not* be compliant with the C++ standard, since the call
to tex1Dfetch would only be omitted at the mercy of the compiler's
optimizer.