gpu/potentials: add method rr_cut() to potential functor
This is needed to smooth the potentials at the cutoff distance.

rr_cut() returns the *square* of the cutoff distance, since this is the
parameter that is used to compare the distance of two particles to the
cutoff distance.

Further note that these methods are only a temporary solution. In the
future, we should move the cutoff distances (for each species pair) to
pair_trunc.