Rewrite reduction on GPU for a simple and flexible interface.
The previous reduction interface required many template parameters, and
yet was inflexible, e.g. incapable of computing the velocity of centre
of mass considering non-unit masses stored as the fourth component of
gpu::particle::velocity().

Rewrite the reduction class and kernel wrapper using a functor approach.
A reduction functor resembles an accumulator that stores a value (or
multiple values, e.g. momentum sum and mass sum for the velocity of
centre mass). The functor has two operator() methods, one accumulates
a value from the input array, the other accumulates another accumulator.

Rewrite the __device__ function reduce to reserve the shared memory
array. For kernels that include reduction besides other calculations,
only an accumulator is required, e.g. reduce<threads>(acc).

Calculate the maximum number of threads for the reduction kernel
using template meta-programming, based on the size of the array
element type and the maximum size of the shared memory. A single
instantiation of reduce_kernel<accumulator_type> automatically
generates kernels for 32 threads up to the maximum number of
allowed threads per block, as a power of two.

Note that we still lack runtime configuration of the number of blocks
and threads, which will require detailed benchmarks for each element
type and CUDA device.

In thermodynamics, add functors for kinetic energy and velocity of
centre of mass, considering the recently added per-particle masses.

Reuse kinetic_energy functor in gpu::integrators::verlet_nvt_hoover.