algorithm/gpu/reduce: avoid DefaultConstructible requirement on accumulator_type
Regardless of whether cuda::host::allocator or halmd::raw_allocator is
used, std::vector<T> requires T to be DefaultConstructible if the vector
is constructed given an initial size.

Use reserve() instead to avoid DefaultConstructible requirement on
accumulator_type. This also avoids a needless initialisation of the
block accumulators in host memory, though this initialisation is
negligible performance-wise given the small number of blocks.