test: fix global fixture ctest_full_output when linking with -Wl,--as-needed
When linking a unit test with -Wl,--as-needed, the test_tools_ctest
library will be dropped by the linker, since its symbols are not
actually used by the test.

To allow linking with -Wl,--as-needed, move the global fixture to
a header file <test/tools/ctest.hpp>, which will be included by all
unit tests. This creates a strong dependency on the test_tools_ctest
library.