test: #include <test/tools/ctest.hpp> in all unit tests
This includes a global fixture which prints CTEST_FULL_OUTPUT to avoid
truncation of test output by ctest, prints the HALMD version, and
enables logging to console.

The header must be included since commit 415ac8d, to create an explicit
dependency on the library in ${HALMD_TEST_LIBRARIES} that contains the
global fixture, and thus allow linking with -Wl,--as-needed.