test: cleanup link libraries of unit tests
Remove ${HALMD_COMMON_LIBRARIES}, which is included in ${HALMD_TEST_LIBRARIES}.

Avoid linking against libhalmd.so, and use direct dependencies on HALMD
static libraries. The linking of unit tests against libhalmd.so is
convenient, but requires that *every* module needs to be compiled
in order to link a single unit test.