cmake: default to -arch=compute_12 -code=compute_12,sm_13,sm_20 for CUDA code
Compile for CUDA compute version 1.2, and generate PTX 1.2 code
as well as binary code for targets of compute capability 1.3 and 2.0.
This ensures backward and forward compatibility with targets of
compute capability ≥ 1.2 through JIT compilation at program
startup, while providing binary code for Tesla and Fermi GPUs.

The additional sm_13 target adds 1 Mb (from 10 Mb) to halmd/gpu.so.

Note that we compile with -arch=compute_12 to disable native double
precision (present with compute_13), and to avoid performance penalty
of IEEE-compliant floating-point (present with compute_20).

We will raise the default compute version later to enable native double
precision, as an alternative to double-single precision, and when all
kernels are optimised for Fermi GPUs.