particle: support get and set of particle arrays with generic iterators
By suggestion of Felix, replace non-template get and set methods
accepting an STL vector with template methods accepting any forward
iterator (for get method) or pair of iterators (for set method).

This allows use of other container types in C++ modules, while
retaining the convenience of the Lua table to STL vector converter
for Lua modules and scripts.

Note the simplified use of set methods in the particle unit tests,
where we simply pass a counting_iterator or lattice_iterator.

This commit also fixes const-correctness of the get methods.