test: test performance of multi_range_for_each on host
While the goal of multi_range_for_each is to allow iteration over a
multi-dimensional range of arbitrary dimensions, which is not possible
with nested for loops, this test ensures that multi_range_for_each is
equivalent to nested for loops in terms of performance.

Register parametrised test cases without BOOST_PARAM_TEST_CASE, since
the test cases should be interleaved, to allow a direct comparison of
multi_range_for_each and nested for loops for a particular array size.
Manually register test cases using init_unit_test() function, which
requires BOOST_TEST_MODULE not to be set.