gpu/binning: add method get_cell() to copy cells to multi-range output iterator
A multi-range iterator is a functor that accepts a multi-dimensional
index of array type, and returns an output iterator for the given
index. The particle indices in the cell of the given index are
then copied to the returned output iterator.

This method will be used for a unit test of particle binning.