cuda_wrapper: support CUDA function attributes in CUDA ≥ 4.1
CUDA 4.1 introduced a function cudaFuncGetAttributes() to retrieve
attributes of a device kernel, e.g. the maximum number of threads
supported by the function and the current device.

The CUDA driver API has had a similar function cuFuncGetAttribute()
since CUDA 2.2, however this function requires a parameter of type
CUfunction, which the CUDA runtime API does not seem to provide.