cmake: add -pedantic -Wno-long-long to default GNU C++ compiler flags
This enables further warnings for non-standard-conform code.

The flag -Wno-long-long is needed to disable warnings about long long
integers, which are used in the CUDA headers but not defined in the
C++98 standard.

While the -pendantic flag also enables less useful warnings such as
duplicate semicolons after BOOST_GLOBAL_FIXTURE (which includes a
semicolon itself), it reveals a programming error in the h5xx
library with regard to scalar datasets.