lattice: determine particles per unit cell from lattice primitive
The number of particles per unit cell depends on the type of lattice
and the dimension (1 for primitive lattice, 2 for two-dimensional
close-packed lattice, and 4 for three-dimensional close-packed
lattice).

This parameter may be determined by creating a lattice of 1 unit
cell, and querying the number of particles with lattice::size().

Use a template parameter lattice_type for the kernel wrapper, and define
a single cuda::function for this lattice type. This way the lattice type
is defined by the host part of the module, gpu::lattice.