host/binning: allow less than 3 cells per dimension
There is no physical reason to restrict the lower bound on the number of
cells per dimension. This may even be a valid use case, e.g. in the case
of a slab geometry with one or two cells in two of the dimensions and
many cells in one dimension.

This fixes the unit test of binning on the host.