binning: avoid use of lambda function in test_non_uniform_density
With C++98, we would boost::bind the test function with the test
parameters and register the resulting functor as a test case.

With C++11, use a simple test function (with return type void) as well,
and substitute boost::bind with a lambda function to create the test
case. This is easy to follow, and allows adding small statements to
the test, such as set_cuda_device.

Replace callback0<> with auto, since the function type of the variable
is obvious, while the type boost::unit_test::callback0<> is specific
to Boost.Test.