Build with -static-libstdc++ -static-libgcc if HALMD_USE_STATIC_LIBS=1
For supported compilers on Linux, add -static-libstdc++ -static-libgcc
to default linker flags. This adds support for building a HALMD binary
which depends on the (optional) CUDA libraries, and system libraries
with a stable ABI, only.

Note that this requires CMake 2.8.8, and GCC ≥ 4.5 or Clang ≥ 3.1.

CMake 2.8.8 is needed for CMAKE_CXX_COMPILER_VERSION.

This commit fixes #89.