cmake: generate HALMD archive
This set of rule crafts a tarball of the halmd repository, including the h5xx
submodule and generated documentation as plain-text ReST, HTML with images,
PDF and manual page.

The tarball is created with git archive of the halmd repository, and further
tarballs of submodules, and the documentation, are appended using GNU tar.
This procedure ensures that the commit id of the HALMD commit is preserved,
and may later be queried with

bzip2 -d < halmd-<version>.tar.bz2 | git get-tar-commit-id

If any command fails, the entire chain of commands is repeated,
to ensure that all source and binary files are up-to-date.

Caveats: Before creating an archive, ensure that all submodules are
at the correct commit (as defined in the halmd commit), and purge
any previously built documentation with make -C doc clean.

This commit fixes #88.