add locally smooth cutoff for truncated potentials
A smooth potential cutoff was already implemented in version 0.1. The commit
backports this functionality from the master branch, bundling several commits
contributed by Nicolas Höft and Peter Colberg.

It provides a template functor to the pair_trunc algorithm and implements two
different truncation schemes in mdsim::forces::trunc:

The module 'discontinuous' does nothing and yields a sharp cutoff as before.

The module 'localr4' implements a smoothing function (suggested by J. Horbach)
and modifies the potential locally at the cutoff, in order to ensure C²
continuity of the potential. The impact of the smoothing depends on the
smoothing parameter h.

On the Lua side, the smoothing parameter is managed by each potential and
retrieved by pair_trunc to construct the smoothing module. This solution is not
very nice, but at least consistent with passing the cutoff parameters to the
potentials as well. A better implementation is provided in the master branch.