Include Luabind as a git submodule.
Build static library libluabind.a of Luabind sources.

With the transition from C++98 to C++11, programs linking against Luabind will
abort with a segmentation fault, unless both Luabind and the program are built
with C++11 standard mode and with the same compiler, e.g. GCC 4.7, or Clang 3.1
using libstdc++-4.7.

This C++ ABI incompatibility does not seem to be an issue with Boost libraries,
e.g. Boost may be compiled with GCC 4.4 and C++98, while a program linking
against Boost libraries is compiled with GCC 4.7, or Clang 3.1, and C++11.