gpu/particle: remove argument threads from constructor
The number of threads should be a per-module setting, which we could
implement per kernel using an occupancy calculator or other heuristics.

The particle arrays should support any multiple of 32 threads in terms
of the allocated memory, which will be implemented in the future.

For now, set the default number of threads to 128.