packages.mk: build with --layout=tagged to override system Boost libraries
This adds a suffix -mt to library names, which is preferred by the CMake
module FindBoost over libraries without suffix -mt. This ensures that
the self-compiled Boost libraries are always preferred over libraries
shipped with the distribution.

This resolves the issue worked around in commit:f83555d.

Add flags threading=multi variant=release as a safeguard.