cmake: add CMake flag to forcibly enable or disable GPU support
By default, GPU support is enabled or disabled depending on whether CUDA
is available. If HALMD_WITH_GPU is explicitly set to TRUE, CMake will
fail if CUDA is not available. If HALMD_WITH_GPU is explicitly set to
FALSE, GPU support will be disabled even if CUDA is available.