cmake: fix extraction of module documentation with HALMD_DOC_ONLY
The extraction of inline documentation requires the Lua interpreter,
however the Lua package was not loaded if HALMD_DOC_ONLY was TRUE.

Split the Lua package into LuaLibs and LuaInterp, following the example
of PythonLibs and PythonInterp, and require LuaLibs for compilation and
LuaInterp for documentation.

This fixes a regression caused by commit:d1db13d.