sampler: support multiple runs and revise Lua module
Remove constructor parameters steps and time, and instead use number of
steps passed to sampler::run(). This allows for an arbitrary number of
simulation runs.

Move function to gracefully abort program from posix_signal to sampler,
since this functionality is strongly tied to the integration loop.

Make runtime_estimate usable in the HALMD script, to allow multiple
simulation runs with a single runtime estimate. Parse constructor
parameters first, interval and sample instead of hard-coding these
values.

Remove typedef slot_function_type in sampler, and use boost::function
with explicit signatures instead to increases readability of the code.

Strictly sample with given interval for prepare and sample slots, and
not at the end of the sample run. This avoids surprises in H5MD output
files if the number of steps is not a multiple of the interval.

Remove unneeded slot wrapping of sampler::setup() and sampler::run().

Briefly document sampler methods.

Add stub documentation for posix_signal, runtime_estimate, timer_service.