sampler: move increment of step from core to sampler
Incrementing the simulation step in core::mdstep() is inherently
inconsistent. While the step does not change for prepare and sample
slots in sampler::setup(), i.e. step is zero for both, the step
is different for prepare and sample slots in sampler::run().

The step must be incremented before prepare slots are called,
and thus belongs into the integration loop in sampler::run().