device: support automatic GPU selection using CUDA runtime API
The previous implementation was a work around for a broken nvlock.

The new implementation depends only on the CUDA runtime library,
and avoids an unnecessary dependency on the CUDA driver library.
Further it relieves HALMD of the burden of device management.

This requires nvlock commit 1011490 or later.

nvlock supports GPU selection without modifications to a program:

http://git.colberg.org/nvcuda-tools.git/